% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addINFO_to_table.R
\name{addINFO_to_table}
\alias{addINFO_to_table}
\title{Add the information from featINFO to a table.}
\usage{
addINFO_to_table(
  df1,
  colfeat_df1,
  dfINFO,
  colfeat_dfINFO,
  add_char_to_INFO = FALSE,
  char_to_add_to_INFO = "_INFO"
)
}
\arguments{
\item{df1}{dataframe. The table containing some results (e.g. from statistical analyses) for each feature.}

\item{colfeat_df1}{character of length 1. It is the name of the column in df1 that contains the unique names of features.}

\item{dfINFO}{dataframe. The featINFO table.}

\item{colfeat_dfINFO}{character of length 1. It is the name of the column in dfINFO that contains the unique names of features.}

\item{add_char_to_INFO}{logical. Shoudl a string of character to be added to each column from the dfINFO?}

\item{char_to_add_to_INFO}{character of length 1. If add_char_to_INFO is TRUE, pass here the string of character that will be added to each column from dfINFO.}
}
\value{
A tibble. The combined table containing also the information on each feature.
}
\description{
Given a table containing data related to features (for example the output of a statistical analyses) and the related featINFO table, it combines everything in a single table.
}
