% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_scale.R
\name{auto_scale}
\alias{auto_scale}
\title{Auto-scale values}
\usage{
auto_scale(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{logical. Should missing value be stripped before the computation? (Same as for mean and sd)}
}
\value{
a numeric vector with the scaled values.
}
\description{
From a vector of numbers, to each element it subtracts the mean and divide by the standard deviation.
}
