% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_long_vector_of_colors.R
\name{build_long_vector_of_colors}
\alias{build_long_vector_of_colors}
\title{Build a vector with colors.}
\usage{
build_long_vector_of_colors()
}
\value{
A character vectors with 112 different colors.
}
\description{
It creates a vector containing 112 different colors from the pals package; in particular from the following palettes, in this order: trubetskoy, glasbey, kelly, polychrome.
}
