% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_fix_names_needed.R
\name{check_if_fix_names_needed}
\alias{check_if_fix_names_needed}
\title{Check if names can be fixed}
\usage{
check_if_fix_names_needed(x)
}
\arguments{
\item{x}{a character vector.}
}
\value{
logical vector, indicating which element contains special characters or start with a number
}
\description{
It checks if names contain spaces or special characters or if start with number
}
