% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_df_examples.R
\name{create_df_examples}
\alias{create_df_examples}
\title{Create example datasets}
\usage{
create_df_examples(
  datasets = c("all", "df_example", "df_example_targeted", "df_example_targeted_legend",
    "df_example_targeted_compounds_legend", "grad_example", "features_data_example",
    "featinfo_example", "melecule_to_search_example")
)
}
\arguments{
\item{datasets}{one of the following: "all", "df_example", "df_example_targeted", "df_example_targeted_legend", "df_example_targeted_compounds_legend", "grad_example", "features_data_example", "featinfo_example", "melecule_to_search_example". If "all", all of the following dataframes (tibble) will be created as different objects; df_example, generic dataframe useful for testing most statistical functions; df_example_targeted, example for the first argument of get_targeted_elaboration; df_example_targeted_legend, example for the second argument of get_targeted_elaboration; df_example_targeted_compounds_legend, example for the third argument of get_targeted_elaboration; grad_example, example for get_phase_amount; features_data_example, example of a feature data table; featinfo_example, example of a featINFO table; melecule_to_search_example, example of known standards to pass to the third argument of checkmolecules_in_feat_table.}
}
\value{
it directly creates the object(s) in the current environment.
}
\description{
It create mock datasets that can be used to test other functions of this package.
}
