% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutP.R
\name{cutP}
\alias{cutP}
\title{Cutting the P-value if < 0.001}
\usage{
cutP(x)
}
\arguments{
\item{x}{a single number, which must be between 0 and 1.}
}
\value{
a single character
}
\description{
Returns "< 0.001" if a number is less than 0.001.
}
