% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_FELLA_enrichment_analysis.R
\name{do_FELLA_enrichment_analysis}
\alias{do_FELLA_enrichment_analysis}
\title{Do FELLA enrichment_analysis.}
\usage{
do_FELLA_enrichment_analysis(
  organism_code,
  KEGG_codes,
  path_databases = "C:/databases/FELLA/",
  output_prefix = "",
  output_suffix = ""
)
}
\arguments{
\item{organism_code}{character of length 1. The KEGG organism code. For example, for homo sapiens (human) is 'hsa'. All codes are here: https://www.genome.jp/kegg/tables/br08606.html.}

\item{KEGG_codes}{Character. KEGG codes of compounds.}

\item{path_databases}{Character of length 1. Path of local organism databases.}

\item{output_prefix}{Character of length 1. What to add at the beginning of generated file names and object names.}

\item{output_suffix}{Character of length 1. What to add at the end of generated file names and object names.}
}
\value{
Export csv table and png file, along with creating object in the global environment.
}
\description{
Given KEGG codes and a defined organism, it performs the enrichment_analysis with the FELLA package.
}
