% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_names.R
\name{fix_names}
\alias{fix_names}
\title{Fix names}
\usage{
fix_names(x, if_start_with_number_add = "m")
}
\arguments{
\item{x}{a character vector.}

\item{if_start_with_number_add}{character of length 1. If the name start with a number, a character indicated in this argument will be added at the beginning.}
}
\value{
a single character with special character replaced.
}
\description{
It removes special characters from a character vector, useful to modify column names. Spaces and hyphens will be replaced with the underscore,
}
