% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_AnnoLevel_to_ordered_number.R
\name{from_AnnoLevel_to_ordered_number}
\alias{from_AnnoLevel_to_ordered_number}
\title{Get an numeric vector from a factor vector of annotation levels.}
\usage{
from_AnnoLevel_to_ordered_number(AnnoLevel_vector)
}
\arguments{
\item{AnnoLevel_vector}{a factor vector, containing annotation levels.}
}
\value{
A numeric vector ranking the annotation levels.
}
\description{
Given a factor vector of annotation levels, it creates a numeric vector. Useful for ranking.
}
