% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_ordered_number_to_AnnoLevel.R
\name{from_ordered_number_to_AnnoLevel}
\alias{from_ordered_number_to_AnnoLevel}
\title{Get a factor vector of annotation levels from a numeric vector.}
\usage{
from_ordered_number_to_AnnoLevel(
  AnnoLevel_as_ordered_number,
  ident_levels = c("1", "2a", "2b", "3a", "3b", "3c", "4a", "5")
)
}
\arguments{
\item{AnnoLevel_as_ordered_number}{a numeric vector, containing annotation levels converted as numeric.}

\item{ident_levels}{character. The levels of the factor vector to be created.}
}
\value{
A factor vector of annotation levels.
}
\description{
Given a numeric vector, it creates a factor vector of annotation levels. It's the opposite of from_AnnoLevel_to_ordered_number.
}
