% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentab_FC_more_than2levels.R
\name{gentab_FC_more_than2levels}
\alias{gentab_FC_more_than2levels}
\title{Generate a Table with Fold Change even if the factor has more than 2 levels}
\usage{
gentab_FC_more_than2levels(
  df,
  v,
  f,
  second_to_first_ratio = TRUE,
  paired = FALSE,
  are_log_transf = FALSE,
  log_base = 2,
  only_on_positive = FALSE
)
}
\arguments{
\item{df}{a dataframe.}

\item{v}{a character vector. Each element must correspond to a column name of the df, each of which must contain numeric values.}

\item{f}{character vector of length 1. Name of the column of df containing the factor variable considered for performing the Fold Change analysis.}

\item{second_to_first_ratio}{logical. If TRUE the second group/first group ratio will be computed, if FALSE the first group/second group ratio will be computed.}

\item{paired}{logical. If FALSE it performs FC on mean of the two groups. If TRUE it performs FC for each pair and then compute the mean.}

\item{are_log_transf}{logical. If you really need to perform this FC analysis on already log-transformed data, specify here as TRUE, and the subtraction will be performed instead of the ration.}

\item{log_base}{numeric of length 1. Specify here the base of the logarithm to calculate the logFC or, if are_log_transf is TRUE; the base of the logarithm the that were used to transform the data.}

\item{only_on_positive}{logical. If TRUE, FC analysis will be performed only if all values are positive (if paired), or if all means are positive (if not paired). The rest will be left as NA.}
}
\value{
A tibble the results of the Fold Change analysis for each pair of levels of the factor variables.
}
\description{
Given a dataframe and a set of numerical variables of that dataframe, it performs fold changes analysis for each pair of levels of a desired variable and creates a new table.
}
