% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_AnnoLevels_MSDial.R
\name{get_AnnoLevels_MSDial}
\alias{get_AnnoLevels_MSDial}
\title{Get the annotation levels.}
\usage{
get_AnnoLevels_MSDial(feat_info_table_MSDial)
}
\arguments{
\item{feat_info_table_MSDial}{a dataframe. A featINFO table created with get_feat_info_from_MSDial.}
}
\value{
A tibble, the featINFO table with an additional column with the annotation levels.
}
\description{
Given a featINFO table obtained from MSDIAL with the function get_feat_info_from_MSDial, it adds a column with the annotation levels according to https://doi.org/10.1007/s00216-022-04207-z.
}
