% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feat_info_from_MSDial.R
\name{get_feat_info_from_MSDial}
\alias{get_feat_info_from_MSDial}
\title{Get a featINFO table from the MSDIAL export file.}
\usage{
get_feat_info_from_MSDial(MSDIAL_raw_table, add_AnnoLevels = FALSE)
}
\arguments{
\item{MSDIAL_raw_table}{a dataframe. Inport on R the exported table using write_tsv or write_csv, then directly pass it in this argument.}

\item{add_AnnoLevels}{logical. If TRUE, an additional column named "AnnoLevels" will be added, and the annotation levels will be calculated considering the cut-offs reported in https://doi.org/10.1007/s00216-022-04207-z.}
}
\value{
A tibble with the information for each feature.
}
\description{
Given a table exported from MSDIAL (exporting the Area from MS-Dial 5.1.230912), it creates the featINFO table, i.e. the table with the full information on each feature.
}
