% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feat_info_from_patRoon.R
\name{get_feat_info_from_patRoon}
\alias{get_feat_info_from_patRoon}
\title{Get a featINFO table from the patRoon outputs.}
\usage{
get_feat_info_from_patRoon(
  patRoon_featureGroups,
  patRoon_MFsummary = NULL,
  MFsummary_score_columns = c("individualMoNAScore", "score"),
  add_AnnoLevels = FALSE
)
}
\arguments{
\item{patRoon_featureGroups}{dataframe. Features groups and related intensities obtained after XCMS. This table should contain the feature group names in a column named "group", the retention times a column named "ret", the m/z ratio in a column named "mz", and the intensities in all other columns.}

\item{patRoon_MFsummary}{NULL or a dataframe obtained from converting as.data.table the output of the function generateCompounds from patRoon. At least the column named "group" and "score" must be present, and if multiple molecules are annotated for each feature group, the one with the top score as provided in the MFsummary_score_columns argument will be considered.}

\item{MFsummary_score_columns}{character of length 1 or more. The name of the column(s) of patRoon_MFsummary that will be used as prioritization score to assign the candidate. The first provided column will be used at first, if that score is tied, the second provided column will be considered, and so for.}

\item{add_AnnoLevels}{logical. If TRUE, and if a MFsummary table was loaded with suitable individualMoNAScore, an additional column named "AnnoLevels" will be added, and the annotation levels will be calculated considering the cut-offs reported in https://doi.org/10.1007/s00216-022-04207-z.}
}
\value{
A tibble with the information for each feature.
}
\description{
Given the featureGroups, components, and MFsummary tables, it creates the featINFO table, i.e. the table with the full information for each feature.
}
