% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phase_amount.R
\name{get_phase_amount}
\alias{get_phase_amount}
\title{Get the amount of mobile phases needed for a sequence.}
\usage{
get_phase_amount(grad, ambient = 2, an = 1)
}
\arguments{
\item{grad}{a dataframe with 4 columns. The first column contains the minute of each gradient step. The second column is the flow in mL/min of each gradient step. The third column is the percentage of the A-phase at each step. The four column is the percentage of the B-phase at each step.}

\item{ambient}{numeric of length 1. The number of minutes the system stays at the t0 conditions between each analysis.}

\item{an}{numeric of length 1. The number of analyses in the sequence.}
}
\value{
It prints the information on the console.
}
\description{
Given a dataframe containing the gradient steps, it calculates the amount of each phase needed for the sequence.
}
