% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCombined_featTable_MSDial.R
\name{GetCombined_featTable_MSDial}
\alias{GetCombined_featTable_MSDial}
\title{Combining three tables of feature intensities in a combined table.}
\usage{
GetCombined_featTable_MSDial(
  featTable1,
  featINFO1,
  name1 = "table1",
  featTable2,
  featINFO2,
  name2 = "table2",
  featTable3,
  featINFO3,
  name3 = "table3",
  mz_widow = 0.005,
  rt_window = 2
)
}
\arguments{
\item{featTable1}{a dataframe, containing feature intensities.}

\item{featINFO1}{a dataframe, containing the featINFO related to featTable1.}

\item{featTable2}{a dataframe, containing feature intensities.}

\item{featINFO2}{a dataframe, containing the featINFO related to featTable2.}

\item{featTable3}{a dataframe, containing feature intensities.}

\item{featINFO3}{a dataframe, containing the featINFO related to featTable3.}

\item{mz_widow}{numeric of length 1. The mass-to-charge ratio window within which features could be grouped as the same feature.}

\item{rt_window}{numeric of length 1. The retention time window within within features could be grouped as the same feature.}
}
\value{
A tibble. The combined table of feature intensities.
}
\description{
Given three different tables of feature intensities and the related featINFO tables, it combines the feature intensities in a single table.
}
