% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_scale.R
\name{mean_scale}
\alias{mean_scale}
\title{Mean-scale values}
\usage{
mean_scale(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{logical. Should missing value be stripped before the computation? (Same as for mean)}
}
\value{
a numeric vector with the scaled values.
}
\description{
From a vector of numbers, it subtracts the mean to each element.
}
