% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_scale.R
\name{pareto_scale}
\alias{pareto_scale}
\title{Pareto-scale values}
\usage{
pareto_scale(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{logical. Should missing value be stripped before the computation? (Same as for mean and sd)}
}
\value{
a numeric vector with the scaled values.
}
\description{
From a vector of numbers, to each element it subtracts the mean and divide by the squared root of the standard deviation.
}
