% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration_curves.R
\name{plot_calibration_curves}
\alias{plot_calibration_curves}
\title{Plotting calibration curves of a targeted elaboration}
\usage{
plot_calibration_curves(targeted_elaboration)
}
\arguments{
\item{targeted_elaboration}{list, an output of the function get_targeted_elaboration.}
}
\value{
a list with ggplots, one for each linear model computed. You can print it to the console to see the plots, or export them using the function export_figures.
}
\description{
It creates plots of the calibration curves from a targeted elaboration.
}
