% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_scale.R
\name{range_scale}
\alias{range_scale}
\title{Range-scale values}
\usage{
range_scale(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{logical. Should missing value be stripped before the computation? (Same as for mean, max and min)}
}
\value{
a numeric vector with the scaled values.
}
\description{
From a vector of numbers, to each element it subtracts the mean and divide by the difference between the maximum and the minimum values.
}
