% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_the_result.R
\name{round_the_result}
\alias{round_the_result}
\title{Round a number to defined digital places}
\usage{
round_the_result(x)
}
\arguments{
\item{x}{a single number.}
}
\value{
a single number.
}
\description{
If the number is less than 10, it returns it rounded with 3 digits,
if it's less than 100 with 2 digits, if it's less than 1000 with 1 digit, otherwise with 0 digits.
}
