% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality_saphiro_table.R
\name{test_normality_saphiro_table}
\alias{test_normality_saphiro_table}
\title{Test normality with a Shapiro–Wilk test}
\usage{
test_normality_saphiro_table(df, v, pvalcutoff = 0.05, cutpval = FALSE)
}
\arguments{
\item{df}{a dataframe.}

\item{v}{a character vector. Each element must correspond to a column name of the df, each of which must contain numeric values.}

\item{pvalcutoff}{a numeric of length 1, must be between 0 and 1, indicating the p-value cut-off.}

\item{cutpval}{logical. Do you want to cut the P-value using the function cutP of the present package?}
}
\value{
A tibble with a number of rows equal to elements specified in v. For each of those, it reports the results of the Saphiro test and whether those data are normally distributed considering the given p-value cutoff.
}
\description{
Given a dataframe and a set of numerical variables of that dataframe, it tests the normality of those data by performing a Shapiro–Wilk test to each desired variable.
}
