% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_prefixing.R
\name{zero_prefixing}
\alias{zero_prefixing}
\title{Zero prefixing numbers}
\usage{
zero_prefixing(
  numbers,
  highest = max(numbers),
  additional_prefix = NULL,
  additional_suffix = NULL
)
}
\arguments{
\item{numbers}{numeric. Numbers to be converted.}

\item{highest}{numeric of length 1. The highest number, used to define the numbers of zeros. By default, it uses the max of the vector passed to numbers.}

\item{additional_prefix}{NULL or character of length 1. An additional prefix to add at the beginning of each number.}

\item{additional_suffix}{NULL or character of length 1. An additional suffix to add at the beginning of each number.}
}
\value{
A character vector with the added zeros as prefix.
}
\description{
It converts a number to a character with zeros as prefix, useful for having them ordered by name correctly.
}
