% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_duplicated.R
\name{fix_duplicated}
\alias{fix_duplicated}
\title{Fix duplicate names}
\usage{
fix_duplicated(
  x,
  zeros = TRUE,
  define_highest_for_zeros = NULL,
  start_with_zero = FALSE,
  exclude_the_first = TRUE,
  NA_as_character = FALSE
)
}
\arguments{
\item{x}{a character vector.}

\item{zeros}{logical. Should leading zero be added?}

\item{define_highest_for_zeros}{NULL or an numeric integer of length 1. If zeros is TRUE, you can decide the highest number to define the number of zeros. If this argument is NULL, the highest number will be defined based on the highest number of duplicated.}

\item{start_with_zero}{logical. If TRUE, the first number added for solving the duplicated will be 0, if FALSE will be 1.}

\item{exclude_the_first}{logical. If TRUE, the suffix will not be added only from the second duplicated on.}

\item{NA_as_character}{logical. If FALSE, the missing values will not be affected by this function, if TRUE, any missing values duplicated will be considered as a character "NA".}
}
\value{
a single character with duplicated solved, thus with only unique elements.
}
\description{
It adds sequential numbers to solve duplicates in a character vector.
}
