% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_duplicated.R
\name{fix_duplicated}
\alias{fix_duplicated}
\title{Fix duplicate names}
\usage{
fix_duplicated(x, zeros = TRUE, NA_as_character = FALSE)
}
\arguments{
\item{x}{a character vector.}

\item{zeros}{logical. Should leading zero be added?}

\item{NA_as_character}{logical. If FALSE, the missing values will not be affected by this function, if TRUE, any missing values duplicated will be considered as a character "NA".}
}
\value{
a single character with duplicated solved, thus with only unique elements.
}
\description{
It adds sequential numbers to solve duplicates in a character vector.
}
