% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feat_table_from_patRoon.R
\name{get_feat_table_from_patRoon}
\alias{get_feat_table_from_patRoon}
\title{Get a table of feature intensities from the patRoon output.}
\usage{
get_feat_table_from_patRoon(patRoon_featureGroups)
}
\arguments{
\item{patRoon_featureGroups}{a dataframe. Features groups and related intensities obtained after XCMS. This table should contain the feature group names in a column named "group", the retention times a column named "ret", the m/z ratio in a column named "mz", and the intensities in all other columns.}
}
\value{
A tibble with the feature intensities.
}
\description{
Given the featureGroups table obtained from patRoon, it creates a table with only the name of features in the first column and the intensities of features in each other column.
}
