% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoxplots.R
\name{getBoxplots}
\alias{getBoxplots}
\title{Create some simple boxplots.}
\usage{
getBoxplots(df, v, f)
}
\arguments{
\item{df}{dataframe. Containing data to plot.}

\item{v}{character. The names of the columns of df containing numerical data to plot}

\item{f}{character of length 1. The names of the factor variable used to group the boxplots.}
}
\value{
A ggplot object (if v is only 1), or a list of ggplots (if v contains 2 or more elements).
}
\description{
Given a table containing data, it create boxplots for each desired variable.
}
