% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPCA.R
\name{getPCA}
\alias{getPCA}
\title{Get Principal components analysis.}
\usage{
getPCA(
  df,
  v,
  s = NULL,
  f = NULL,
  col_pal = NULL,
  labels_on_loading = TRUE,
  PC_to_plot = c("PC1", "PC2")
)
}
\arguments{
\item{df}{dataframe. Containing data to plot.}

\item{v}{character. The names of the columns of df containing numerical data (ideally, they should be already centered and scaled!!).}

\item{s}{NULL or a character of length 1. The name of the column of df containing the sample names. Pass it only if you want sample names on the score plot.}

\item{f}{NULL or character of length 1. The name of the column of df containing the factor variable. Pass it only if you want colored points and ellipses on the score plot.}

\item{col_pal}{a character vector containing colors. If NULL, colors from the pals package will be used (see function build_long_vector_of_colors).}

\item{labels_on_loading}{logical. If TRUE, it will report the names of features the loading plot.}

\item{PC_to_plot}{character of length 2. The principal components to plots on the score and loading plots.}
}
\value{
A list with 4 objects:
\itemize{
\item \code{df_with_scores_table}: the df with additional columns containing the scores.
\item \code{loadings_table}: the tibble containing the loadings.
\item \code{score_plot}: a ggplot object with the score plot.
\item \code{loading_plot}: a ggplot object with the loading plot.
}
}
\description{
Given a table containing data, it performs the Principal components analysis.
}
